requirements.txt
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
cryptography==40.0.1
idna==3.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
pdfminer.six==20221105
platformdirs==3.2.0
pycparser==2.21
PyPDF2==3.0.1
pypiwin32==223
pyttsx3==2.90
pywin32==306
requests==2.28.2
tika==2.6.0
tomli==2.0.1
urllib3==1.26.15
